[Update] 플레이어 이동속도 저장누락
- 이동속도 미 저장되어 게임종료 후 재시작 시 기본값으로 초기화 되는 현상(수정완료)

+++ Assets/Scripts/Save Data.cs
@@ -8,7 +8,7 @@ public class SaveData
     public int currentHp;
     public int attack1;
     public int attack2;
-    public int movespeed;
+    public float movespeed;
     public int skillPoints;
     public int fireballLevel;
     public int fireballDamage;
